Implement Authentication using Google OAuth
This adds authentication against Googles OAuth provider.  With this
implemented it should be straight forward to add other authentication
providers to our App such as ActiveDirectory, OneLogin For Governement
or social logins such as FaceBook.

We need to properly implement user storage and user management in to the
system along with RBAC.