Improved error handling and more joined up journey
This brings in the new source and dimension creation page and functions
and improves the route through the publish journey.  Most requests are
now gets with posts redirecting to the next page in the journey on
success.

Steps have been taken to reduce code reuse and the code now makes much
greater use of sessions to keep the users context on what they are
working on.