Refactor the frontend to support the changes to the backend
The backend refactor resulted in a number of changes to the frontend
mostly around changing `imports` to `fact-tables`.  The update adds
support for line numbers in csv table previews.  It also adds support
for proper GDS style pagination which resulted in the preview and view
pages needing to have their layouts adjusted.  This will need to be
cleared by UX and content.  It does however bring these pages much more
in line with the GDS Design System and service standard.