Implement measures and preview
This PR implements measures and data cube previews in the publisher
journey.  It adds the ability to download the cube in multiple formats
including CSV, Excel and Parquet as well as downloading the raw cube in
duckdb format.

Measures are added as a look up table upload in the measures lookup
table format.

The preview page currently lacks the topics associated with the cube due
to a bug on the backend which doesn't currently pass the topics through
in a user readable format.