Implemetn update journey in the frontend
This commit aligns the frontend with the backend following changes to
how sources are produced and stored.  It gets rid of the idea of a
revision having multiple file imports and lays the ground work for just
working with revisions instead of needing to track which file import is
being worked on.

Additional changes include the new renderes for markdown documents.  The
new cookie statement and a completely linked up footer page.