Fix issues around developer view and fact table issues
Introduces the handling of the bad fact tables.  Fact tables are now
validated on the backend as part of the processing of the column
sources.  This should prevent a situation where a user makes it as far
as creating dimensions and then not being able to proceed with dimension
creation and updates.