Fix uploader bug affecting updates
There was a bug in the the client side code affecting the update
journey.  This was caused by trying to use a server side variable in the
client side JavaScript code.  Adjusted the code to look for the form
element and read the property from there and then adding it to the ajax
request.