Add support for long cube builds
End points which regenerate the cube now return always return a build ID
which we can query with the frontend.  If a build takes more than 10
seconds (up for discussion) a page is rendered to say the cube is
building.  If it takes less than 10 seconds it automatically redirects
to the originally intended page.